Rabinizer integeration branch



See merge request !7

+++ prism/src/parser/PrismParserTokenManager.java
@@ -1059,7 +1059,7 @@ public class PrismParserTokenManager implements PrismParserConstants
 					}
 				} while (i != startsAt);
 			} else {
-				int hiByte = (int) (curChar >> 8);
+				int hiByte = curChar >> 8;
 				int i1 = hiByte >> 6;
 				long l1 = 1L << (hiByte & 077);
 				int i2 = (curChar & 0xff) >> 6;
