+++ prism/src/simulator/method/CIiterations.java
@@ -157,7 +157,7 @@ public final class CIiterations extends CIMethod
 		// 2 iterations needed to compute variance of the sampler
 		if (sampler.getVariance() <= 0.0 || iters < 2)
 			return 0;
-		return 10 * ((int) (100.0 * (double) (iters + 1) / (sampler.getVariance() * squaredQuantile / (width * width))) / 10);
+		return 10 * ((int) (100.0 * (iters + 1) / (sampler.getVariance() * squaredQuantile / (width * width))) / 10);
 	}
 
 	@Override
